Adjust the Image Digest Filter
Previously the code that resolves the image metadata was hard coded to
only look for an image digest if the string containing the image id was
prefixed with "docker-pullable".  This prefix filter had the potential
to ignore many other possible scenarios.

This patch separates the logic for digest parsing to a function, changes
the string to split on "sha" and adds it back to the resulting string and
adds a unit test for the common known cases.

Signed-off-by: Ryan Brady <ryan.brady@anchore.com>