Remove the empty string checks from the integration test
Removed the checks for ClusterName and InventoryType because during
integration testing the GetAppConfig() function relies on the init()
function to generate the config from the config file. However it seems
integration testing simply generates an empty config which leaves the
cluster name and inventory type empty. Leaving the decision to fix for a
later date as implementing a default cluster name could overwrite
inventories in anchore enterprise so it will need some thought and seems
out of scope for this PR.

Signed-off-by: James Petersen <jpetersenames@gmail.com>