Fix Go-Bouncer Installation
The install script for go-bouncer is load-bearing; it will download a
license database that it uses for scanning licenses. The Makefile was
updated to use the install script to install bouncer instead of through
go install.

The versions for goreleaser, golang-ci, and go-bouncer are also now
included as constants at the top of the Makefile to make future
modifications to the versions easier.

Signed-off-by: Dustin Schoenbrun <dustin.schoenbrun@anchore.com>
Co-Authored by: Weston Steimel <weston.steimel@anchore.com>