fix: try to catch invalid responses on Enterprise API redirect
Go's http client will automatically follow redirects, we want to
preserve the ability to follow redirects but we need to catch cases
where the redirect returns a 2XX status but the POST was not actually
successful. A simple check is read the response body (if one is
present) and check if it is valid JSON. If it is not JSON, as is the case
with cloudflare where it redirects to a login screen then we can return
an error and log the response body to give customers an indication of
where the redirect has taken them so it can be remediated at the proxy.

Signed-off-by: Bradley Jones <bradley.jones@anchore.com>