fix: redact passwords under account-routes
This commit makes changes so that all passwords and private keys are
redacted in logs, when json marshalled and when yaml marshalled. It
also fixes so that config keys have correct names (e.g. include '-' for
multi-word keys)) when marhalled to yaml and json. Finally, it fixes so
that the 'default' key/value pairs under
metadata-collection(.nodes|.namespaces|.pods) are properly unmarshalled.

Addresses: ENTERPRISE-4159

Signed-off-by: Bob Melander <bob.melander@anchore.com>