test: update tests to support RWMutex in go 1.24
tests are failing because
```
panic: reflect: call of reflect.Value.Int on zero Value [recovered]
        panic: reflect: call of reflect.Value.Int on zero Value
```

Upon digging in further it seems like some of the internal structure of sync.RWMutex has changed.

 Looking at an empty `&sync.RWMutex{}`
In go 1.22 - `&{{0 0} 0 0 {{} 0} {{} 0}}`
In go 1.24 - `&{{{} {0 0}} 0 0 {{} 0} {{} 0}}`
Notice the extra empty struct at the beginning.

In the tests we use reflect to look at the contents of the mutex to determine if the lock is locked or not however this empty struct is what causes the issue and it should be ignored.

Signed-off-by: Bradley Jones <bradley.jones@anchore.com>